requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
idna==3.3
mysqlclient==2.1.0
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
razorpay==1.2.0
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
twilio==7.4.0
urllib3==1.26.8
yapf==0.32.0
